+++ backend/src/main/java/com/ajemi/backend/service/LikeService.java
@@ -1,5 +1,6 @@
 package com.ajemi.backend.service;
 
+import org.springframework.http.HttpStatus;
 import org.springframework.lang.NonNull;
 import org.springframework.stereotype.Service;
 
@@ -7,10 +8,12 @@ import com.ajemi.backend.entity.Like;
 import com.ajemi.backend.entity.Notification.NotificationType;
 import com.ajemi.backend.entity.Post;
 import com.ajemi.backend.entity.User;
+import com.ajemi.backend.exception.ApiException;
 import com.ajemi.backend.repository.LikeRepository;
 import com.ajemi.backend.repository.PostRepository;
 import com.ajemi.backend.repository.UserRepository;
 
+import jakarta.transaction.Transactional;
 import lombok.RequiredArgsConstructor;
 
 @Service
@@ -21,29 +24,29 @@ public class LikeService {
     private final PostRepository postRepository;
     private final UserRepository userRepository;
     private final NotificationService notificationService;
-
+@Transactional
     public String toggleLike(@NonNull Long postId, String username) {
 
         // 1) نجيبو البوست
         Post post = postRepository.findById(postId)
-                .orElseThrow(() -> new RuntimeException("Post not found"));
-
+               .orElseThrow(() -> new ApiException("Post not found", HttpStatus.NOT_FOUND));
     // 2️⃣ actor (لي دار like)
         User actor = userRepository.findByUsername(username)
-            .orElseThrow(() -> new RuntimeException("User not found"));
+           .orElseThrow(() -> new ApiException("User not found", HttpStatus.NOT_FOUND));
+
+        if (actor.isBanned()) {
+            throw new ApiException("Your account is suspended.", HttpStatus.FORBIDDEN);
+        }
         User receiver = post.getAuthor();
         // 3) نشوفو واش دار like قبل
-        var  existing = likeRepository.findByPostAndUser(post, actor);
+       var existing = likeRepository.findByPostIdAndUserId(postId, actor.getId());
 
         if (existing.isPresent()) {
             // --- كان دير like → نديرو UNLIKE
             likeRepository.delete(existing.get());
             return "unliked";
         }
-        // if (actor.getId().equals(receiver.getId())) {
-        //     return "liked"; // أو return بلا notification
-        // }
-        // --- ما دارش like → نديرو LIKE
+
         Like like = new Like();
         like.setPost(post);
         like.setUser(actor);
@@ -57,10 +60,8 @@ public class LikeService {
         return "liked";
     }
 
-    public int getLikesCount(@NonNull Long postId) {
-        Post post = postRepository.findById(postId)
-                .orElseThrow(() -> new RuntimeException("Post not found"));
+    // public int getLikesCount(@NonNull Long postId) {
 
-        return likeRepository.countByPost(post);
-    }
+    //     return likeRepository.countByPostId(postId);
+    // }
 }
